+++ src/lib.rs
@@ -16,12 +16,24 @@ pub use adaptor::{
         json::JsonParser,
         json5::Json5Parser,
         ron::RonParser,
+        ser::SerParser,
         toml::TomlParser,
         yaml::YamlParser,
     },
-    source::Source,
+    source::{
+        Source,
+        cmd::CmdSource,
+        env::EnvSource,
+        file::FileSource,
+        ser::SerSource,
+        string::StringSource,
+    },
 };
 pub use error::Error;
 pub(crate) use error::Result;
+pub use realme::{
+    Realme,
+    builder::RealmeBuilder,
+};
 pub use utils::Map;
 pub use value::Value;